# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kubeasy CLI API",
    "version": "1.0.0",
    "description": "Public API for the Kubeasy platform consumed by the CLI. Web routes use session cookie authentication; CLI routes use Bearer token authentication."
  },
  "servers": [
    {
      "url": "https://kubeasy.dev",
      "description": "Production"
    },
    {
      "url": "http://localhost:3024",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie set after login via better-auth"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained via `kubeasy login`. Used by the CLI and accepted on all public API routes."
      }
    },
...
  },
  "paths": {
...
    "/api/challenges": {
      "get": {
        "operationId": "listChallenges",
        "summary": "List challenges",
        "tags": ["CLI"],
        "security": [
          {
            "SessionAuth": []
          },
          {
            "BearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["easy", "medium", "hard"]
            },
            "required": false,
            "name": "difficulty",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "theme",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "showCompleted",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "theme": {
                            "type": "string"
                          },
                          "themeSlug": {
                            "type": "string"
                          },
                          "difficulty": {
                            "type": "string",
                            "enum": ["easy", "medium", "hard"]
                          },
                          "type": {
                            "type": "string"
                          },
                          "typeSlug": {
                            "type": "string"
                          },
                          "estimatedTime": {
                            "type": "integer"
                          },
                          "initialSituation": {
                            "type": "string"
                          },
                          "ofTheWeek": {
                            "type": "boolean"
                          },
                          "completedCount": {
                            "type": "integer"
                          },
                          "userStatus": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "slug",
                          "title",
                          "description",
                          "theme",
                          "themeSlug",
                          "difficulty",
                          "type",
                          "typeSlug",
                          "estimatedTime",
                          "initialSituation",
                          "ofTheWeek",
                          "completedCount",
                          "userStatus"
                        ]
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  },
                  "required": ["challenges", "count"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges/{slug}": {
      "get": {
        "operationId": "getChallenge",
        "summary": "Get challenge details",
        "tags": ["CLI"],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "slug": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "theme": {
                          "type": "string"
                        },
                        "themeSlug": {
                          "type": "string"
                        },
                        "difficulty": {
                          "type": "string",
                          "enum": ["easy", "medium", "hard"]
                        },
                        "type": {
                          "type": "string"
                        },
                        "typeSlug": {
                          "type": "string"
                        },
                        "estimatedTime": {
                          "type": "integer"
                        },
                        "initialSituation": {
                          "type": "string"
                        },
                        "ofTheWeek": {
                          "type": "boolean"
                        },
                        "available": {
                          "type": "boolean"
                        },
                        "starterFriendly": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "slug",
                        "title",
                        "description",
                        "theme",
                        "themeSlug",
                        "difficulty",
                        "type",
                        "typeSlug",
                        "estimatedTime",
                        "initialSituation",
                        "ofTheWeek",
                        "available",
                        "starterFriendly"
                      ]
                    }
                  },
                  "required": ["challenge"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges/meta": {
      "get": {
        "operationId": "getChallengeMeta",
        "summary": "Get challenge metadata (themes, types, difficulties)",
        "tags": ["CLI"],
        "responses": {
          "200": {
            "description": "Metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "themes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "logo": {
                            "type": "string"
                          }
                        },
                        "required": ["slug", "name", "description", "logo"]
                      }
                    },
                    "types": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "logo": {
                            "type": "string"
                          }
                        },
                        "required": ["slug", "name", "description", "logo"]
                      }
                    },
                    "difficulties": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": ["themes", "types", "difficulties"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges/{slug}/yaml": {
      "get": {
        "operationId": "getChallengeYaml",
        "summary": "Get challenge.yaml",
        "tags": ["CLI"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Raw YAML",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges/{slug}/manifests": {
      "get": {
        "operationId": "getChallengeManifests",
        "summary": "Get challenge manifests tar.gz",
        "tags": ["CLI"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Binary data",
            "content": {
              "application/gzip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}